+++ src/main/java/com/project/doongdoong/domain/counsel/application/CounselServiceImpl.java
@@ -19,9 +19,9 @@ import com.project.doongdoong.domain.counsel.model.Counsel;
 import com.project.doongdoong.domain.counsel.model.CounselType;
 import com.project.doongdoong.domain.counsel.adapter.out.CounselRepository;
 import com.project.doongdoong.domain.user.exeception.UserNotFoundException;
-import com.project.doongdoong.domain.user.model.SocialIdentifier;
-import com.project.doongdoong.domain.user.model.User;
-import com.project.doongdoong.domain.user.repository.UserRepository;
+import com.project.doongdoong.domain.user.domain.SocialIdentifier;
+import com.project.doongdoong.domain.user.domain.User;
+import com.project.doongdoong.domain.user.application.port.out.UserRepository;
 import com.project.doongdoong.global.dto.response.CounselAiResponse;
 import com.project.doongdoong.global.util.CounselRankingCache;
 import com.project.doongdoong.global.util.WebClientUtil;
