+++ src/main/java/com/project/doongdoong/domain/counsel/model/Counsel.java
@@ -1,7 +1,7 @@
 package com.project.doongdoong.domain.counsel.model;
 
 import com.project.doongdoong.domain.analysis.domain.AnalysisEntity;
-import com.project.doongdoong.domain.user.model.User;
+import com.project.doongdoong.domain.user.domain.User;
 import com.project.doongdoong.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
