+++ src/test/java/com/project/doongdoong/domain/analysis/repository/AnalysisEntityJpaRepositoryTest.java
@@ -10,9 +10,9 @@ import com.project.doongdoong.domain.answer.application.port.out.AnswerJpaReposi
 import com.project.doongdoong.domain.question.domain.Question;
 import com.project.doongdoong.domain.question.domain.QuestionContent;
 import com.project.doongdoong.domain.question.application.port.out.QuestionRepository;
-import com.project.doongdoong.domain.user.model.SocialType;
-import com.project.doongdoong.domain.user.model.User;
-import com.project.doongdoong.domain.user.repository.UserRepository;
+import com.project.doongdoong.domain.user.domain.SocialType;
+import com.project.doongdoong.domain.user.domain.User;
+import com.project.doongdoong.domain.user.application.port.out.UserRepository;
 import com.project.doongdoong.domain.voice.model.Voice;
 import com.project.doongdoong.domain.voice.repository.VoiceRepository;
 import org.junit.jupiter.api.DisplayName;
