+++ src/test/java/com/project/doongdoong/module/ControllerTestSupport.java
@@ -7,8 +7,8 @@ import com.project.doongdoong.domain.answer.application.port.in.AnswerService;
 import com.project.doongdoong.domain.counsel.adapter.in.web.CounselController;
 import com.project.doongdoong.domain.counsel.application.port.in.CounselService;
 import com.project.doongdoong.domain.counsel.application.port.in.CounselStatisticsService;
-import com.project.doongdoong.domain.user.controller.UserController;
-import com.project.doongdoong.domain.user.service.UserService;
+import com.project.doongdoong.domain.user.adapter.in.web.UserController;
+import com.project.doongdoong.domain.user.application.port.in.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
